feat: Write/retrieve chunks using postgres (#17)
* feat: Write/retrieve chunks using postgres

This removes the dependency on Redis, and makes the chunks/embeddings
in the postgres database work.

There are some issues to be addressed, specifically deduplicating cases
where multiple embeddings of the same chunk are retrieved. I plan to
work on those in a follow-up PR, so that we can get the bulk of this in
first.