feat: Code to apply migrations only once (#25)
Uses a migration table to track which migrations have been applied.
Currently assumes that migrations are applied from a single process
(which seems OK given in a multi-node environment we'd likely want to
apply migrations during release, not as part of node startup). Reports
cases where the migration that was applied doesn't match the migration
that was requested (currently doesn't do anything to resolve that case).

This closes #24.