ci: Initial release workflows
These are currently configured to publish `dewy` and `dewy-client` to
test-pypi so that I can overwrite previous versions to verify it works.

Once working, I can switch to `pypi` and/or add that as a separate step.
We could actually run some tests against the package in `test-pypi`
before publishing the artifcats to `pypi`.