Use a string-typed ID for collections, and eliminate the "name" field
Working with collections is difficult. Most API endpoints currently
expect an integer identifier, which the user must find by making an API
query.

The point of collections is to provide logical separation between
documents, so it's more natural use use descriptive names to identify
the collection, ie "cooking-docs" or "prod-env".

This change makes the ID string-typed and user-provided, and drops the
"name" field which is now redundant. This is a breaking change - the SQL
schema is affected as well as the API types. In theory the "name" field
could be promoted to be a new idnetifier through a SQL migration (since
it has a unique constraint), but this seems like overkill at this stage.