feat!: Use taskiq for durable ingestion (#107)
* feat!: Use taskiq for durable ingestion

This moves the core logic into a `domain` module containing the
ingestion and database interactions. The FastAPI routes are in
`backend` and simply translate the incoming requests into calls to
the `domain`. The tasks are defined in `tasks`, and again translate
the task message into calls to the `domain`.

This was necessary because tasks need to use `TaskiqDepends` to inject
things from the worker and API methods need to use `Depends` to inject
things. Thus, we couldn't share dependencies between these methods. It
was thus easier to have the logic do *no* dependency injection, and
have thin layers that do injection / translation on top of that.

- Aligns list_chunks and retrieve_chunks response object
- Allows multiple document IDs in list and retrieve chunk domain
  Not yet exposed / tested.
- Allows retrieval to use offset/limit for paging. Not yet tested.

Future Improvements

- This opens the door to more testing of the domain logic by direct
  invocation / operation on the DB.
- Split ingestion into a sequence of tasks, so that we can commit
  progress, and/or allow for re-entrant ingestion. For instance, if
  we detect that we already created chunks for a document, then we
  should look to fill in missing embeddings, etc.

Breaking Changes:

- Uses `TextChunk` and `ImageChunk` for both retrieval and list chunks.
- Uses `Collection` model directly rather than separate `CreateCollection`.

This partly addresses #75.

* ruff

* fix frontend

* update openapi client

* verify ui and fix some things

* ruff