fix: handle mimetypes with encodings
As reported in #112 and demonstrated in the test (prior to fixing) we
didn't handle mimetypes correctly if they contained an encoding. For
instance `text/html; charset=utf-8`. The fix is to split the string on
the semicolon, and only take the mimetype (appearing before the first
semicolon) for the matching.

This closes #112.